Code refactoring and spotless apply changes
