+++ src/main/java/com/commercetools/project/sync/model/ProductSyncCustomRequest.java
@@ -2,6 +2,10 @@ package com.commercetools.project.sync.model;
 
 import static java.lang.String.format;
 
+import com.commercetools.project.sync.exception.CliException;
+import com.fasterxml.jackson.databind.ObjectMapper;
+import java.io.IOException;
+
 public final class ProductSyncCustomRequest {
 
   private Long fetchSize;
@@ -25,5 +29,20 @@ public final class ProductSyncCustomRequest {
 
   public void setCustomQuery(String customQuery) {
     this.customQuery = customQuery;
+  }
+
+  public static ProductSyncCustomRequest parseProductQueryParametersOption(String customRequest) {
+
+    final ObjectMapper objectMapper = new ObjectMapper();
+
+    final ProductSyncCustomRequest productSyncCustomRequest;
+    try {
+      productSyncCustomRequest =
+          objectMapper.readValue(customRequest, ProductSyncCustomRequest.class);
+    } catch (IOException | IllegalArgumentException e) {
+      throw new CliException(e.getMessage());
+    }
+
+    return productSyncCustomRequest;
   }
 }
